Factor Linux specific functionality out of libzutil
